wave-extra/wave-cookies/support-cookies-deletion.https.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <title>Support deleting a cookie</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="resources/cookies-helper.js"></script>
</head>
<body>
  <div id=log></div>
  <script>
    
    const TEST_CASE = {file: "deletion_test", value: "deletion_test"};
    setCookie(TEST_CASE.file, TEST_CASE.value, 1);
    deleteCookie(TEST_CASE.file, "/");
    test(function () {
      const cookieValue = getCookie(TEST_CASE.file);
      assert_equals(cookieValue, null);
    }, "Support deleting cookies");
    done();

  </script>
</script>
</body>
</html>
